Splits up external parsers from Needle core.
